[Feat/#62] modify IDCardConfirmEditView
